Clean Up for Release V1

+++ Aim Trainer/Assets/Scripts/Mode 2 Scripts/TargetMoving.cs
@@ -35,8 +35,8 @@ public class TargetMoving : MonoBehaviour
 
     void Start()
     {
-        rotationSpeed = BotDifficulty.botRotationSpeed;
-        speed = BotDifficulty.botSpeed;
+        rotationSpeed = (float)BotDifficulty.botRotationSpeed;
+        speed = (float)BotDifficulty.botSpeed;
     }
